(feat) moved Routes.js into components for easier testing
