+++ src/public/components/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import Routes from '../Routes';
+import Routes from './Routes';
 import Navigation from './Navigation';
 import { get } from './AxiosRoutes';
 import { BrowserRouter as Router, Route, Link, Redirect } from 'react-router-dom';
